:art: separate ProductListPage into presentation and container components
