+++ src/shared/nt-core/App.react.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
 
-import ProductListPage from '../nt-products/ProductListPage.react'
+import ProductListPageContainer from '../nt-products/ProductListPage.container.react'
 import MainPageContainer from './MainPage.container.react'
 import NoMatch from './NoMatch.react'
 
 const App = () => (
   <Switch>
     <Redirect exact path='/' to={{ pathname: '/products' }} />
-    <Route exact path='/products' component={ProductListPage} />
+    <Route exact path='/products' component={ProductListPageContainer} />
     <Route path='/products/:productId' component={MainPageContainer} />
     <Route component={NoMatch} />
   </Switch>
